% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciiu.R
\docType{data}
\name{ciiu}
\alias{ciiu}
\title{CIIU}
\format{
Data frame con variables:
\describe{
\item{seccion}{Sección Económica}
\item{division}{División Económica}
\item{grupo}{Grupo Económico}
\item{clase}{Clase Económica}
\item{subclase}{Subclase Económica}
\item{glosa}{Glosa}
}
}
\source{
\url{https://www.ine.cl/}
}
\usage{
ciiu
}
\description{
Clasificador Internacional Industrial Uniforme CIIU 2012 ajustado a Chile
}
\examples{

}
\keyword{datasets}
