% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciiu_digitos.R
\name{ciiu_digitos}
\alias{ciiu_digitos}
\title{Glosas según número de dígitos}
\usage{
ciiu_digitos(digitos = NA_real_)
}
\arguments{
\item{digitos}{Indicar el número de desglose de CIIU 2012}
}
\value{
Una tibble con las glosas según desglose solicitado
}
\description{
La función retorna todos los elementos contenidos en la desagregación
consultada.
}
