% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciiu_glosa.R
\name{ciiu_glosa}
\alias{ciiu_glosa}
\title{Buscar glosas CIIU 2012 con palabras claves}
\usage{
ciiu_glosa(
  let = NA_character_,
  div = NA_character_,
  gru = NA_character_,
  cla = NA_character_,
  sub = NA_character_,
  buscar = ""
)
}
\arguments{
\item{let}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
La función retorna todos los elementos contenidos en la desagregación
consultada.
}
