% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciuo.R
\docType{data}
\name{ciuo}
\alias{ciuo}
\title{CIUO}
\format{
Data frame con variables:
\describe{
\item{gran_grupo}{Gran grupo}
}
}
\source{
\url{https://www.ine.cl/}
}
\usage{
ciuo
}
\description{
Clasificador Internacional Uniforme de Ocupaciones
}
\examples{

}
\keyword{datasets}
