% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciuo_digitos.R
\name{ciuo_digitos}
\alias{ciuo_digitos}
\title{Glosas según número de dígitos}
\usage{
ciuo_digitos(digitos = NA_real_)
}
\arguments{
\item{digitos}{Indicar el número de desglose de CIUO}
}
\value{
Una tibble con las glosas según desglose solicitado
}
\description{
La función retorna todos los elementos contenidos en la desagregación
consultada. Con valor "iricmo" en argumento digitos retorna glosas IRICMO.
}
\examples{
ciuo_digitos()
ciuo_digitos("iricmo")
ciuo_digitos(1)
ciuo_digitos(2)
ciuo_digitos(3)
ciuo_digitos(4)
ciuo_digitos(5)
}
