% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciiu_desagrega.R
\name{clase}
\alias{clase}
\title{Buscar desagregaciones CIIU 2012 dentro de letras}
\usage{
clase(letra = NA_character_)
}
\arguments{
\item{letra}{Sección a desagregar. Como carácter.}
}
\value{
Tabla con número de la desagregación buscada y glosa
}
\description{
Cuatro funciones que retornas divisiones, grupos, clases o subclases dentro de una sección o letra de CIIU 2012.cl
}
\examples{
clase("C")
}
