% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_meses.R
\name{ene3_to_n_meses}
\alias{ene3_to_n_meses}
\title{Funciones para lidiar con meses}
\usage{
ene3_to_n_meses(x, y)
}
\arguments{
\item{let}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
La función retorna todos los elementos contenidos en la desagregación
consultada.
}
