% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciiu_desagrega.R
\name{seccion}
\alias{seccion}
\title{Buscar desagregaciones CIIU 2012 dentro de letras}
\usage{
seccion(letra = NA_character_)
}
\arguments{
\item{letra}{Sección a desagregar. Como carácter.}
}
\value{
Tabla con número de la desagregación buscada y glosa
}
\description{
Entrega glosa de la seccion
}
\examples{
seccion("C")
}
