% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciiu_consulta.R
\name{ciiu_consulta}
\alias{ciiu_consulta}
\title{Consultar CIIU 2012}
\usage{
ciiu_consulta(
  let = NA_character_,
  div = NA_character_,
  gru = NA_character_,
  cla = NA_character_,
  sub = NA_character_
)
}
\value{
A matrix of the infile
}
\description{
La función retorna todos los elementos contenidos en la desagregación
consultada. Se puede filtrar por letra, división, grupo, clase o subclase
}
