% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciiu_glosa.R
\name{ciiu_glosa}
\alias{ciiu_glosa}
\title{Buscar glosas CIIU 2012 con palabras claves}
\usage{
ciiu_glosa(
  let = NA_character_,
  div = NA_character_,
  gru = NA_character_,
  cla = NA_character_,
  sub = NA_character_,
  buscar = ""
)
}
\arguments{
\item{buscar}{Palabra a buscar en glosas}
}
\value{
tablas con todas las desagregaciones cuyas glosas coinciden con lo buscado
}
\description{
Buscar glosas CIIU 2012 con palabras claves
}
\examples{
ciiu_glosa(buscar="Agro")
ciiu_glosa(let="G",buscar="Agro")
}
