% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciiu_desagrega.R
\name{division}
\alias{division}
\title{Consultar CIIU 2012}
\usage{
division(let = NA_character_)
}
\arguments{
\item{let}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
La función retorna todos los elementos contenidos en la desagregación
consultada.
}
